[Polymorphous] IPDPS 2025 Open Source
To build:
1) Please download cloog-isl.tar.gz from https://github.com/kaitingwang/polymorphous-ipdps2025-artifacts.git
and place it inside the llvm-project directory.  Please note both an ARM64 and an X86 binaries have been provided.
The easiest way to build cloog-isl from source is to download and build pluto-0.12.0 and simply copy the
cloog-isl directory over.  Polymorphous shares the same cloog-isl as pluto-0.12.0.

2) Update mlir/lib/Dialect/Affine/Transforms/CMakeLists.txt and mlir/ib/Dialect/Affine/Utils/CMakeLists.txt to
point to your cloog-isl directory and binaries.

3) To build mlir:
  - cd llvm-project
  - mkdir build
  - cmake -G "Unix Makefiles" ../llvm -DLLVM_ENABLE_PROJECTS="mlir" -DLLVM_TARGETS_TO_BUILD=host -DCMAKE_BUILD_TYPE=Release -DLLVM_ENABLE_ASSERTIONS=ON
  - make -j4 mlir-opt
  - make -j4 mlir-affine-validator

4) All tests are kept in mlir/test/Polymorphous.  You may refer to the readme file inside to run Polybench testcases.